DRY up component discovery code in tests
