+++ tests/utils/govuk-frontend-diff.js
@@ -2,8 +2,7 @@ const nunjucks = require('nunjucks');
 const ent = require('ent');
 const prettyhtml = require('@starptech/prettyhtml');
 const { HtmlDiffer } = require('@markedjs/html-differ');
-const glob = require('glob');
-const path = require('path');
+const getGovukComponentList = require('../../scripts/get-govuk-component-list');
 const examples = require('../../.cache/govuk-frontend-examples/all.json');
 
 const htmlDiffer = new HtmlDiffer({
@@ -17,19 +16,8 @@ function cleanHtml(dirtyHtml) {
   }).contents;
 }
 
-// Grab list of components from govuk-frontend
-const govukComponentPath = path.join(
-  path.dirname(require.resolve('govuk-frontend/package.json')),
-  'govuk/components'
-);
-const components = glob
-  .sync(path.join(govukComponentPath, '**/macro.njk'))
-  .map((componentPath) =>
-    path.relative(govukComponentPath, path.dirname(componentPath))
-  );
-
 function diffComponentAgainstReferenceNunjucks(renderCallback) {
-  components.forEach((component) => {
+  getGovukComponentList().forEach((component) => {
     describe(component, () => {
       examples[component].examples.forEach((example) => {
         describe(`${example.name}`, () => {
